requirements.txt
asgiref==3.3.2
attrs==22.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
celery==5.3.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
defusedxml==0.7.0rc1
Django==4.2.10
django-cors-headers==4.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
django-oauth-toolkit==2.3.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
drf-yasg==1.21.7
flake8==7.0.0
flake8-bugbear==23.3.23
flake8-builtins==2.1.0
flake8-commas==2.1.0
flake8-import-order==0.18.2
flake8-variables-names==0.0.5
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.7.0
oauthlib==3.1.0
pep8-naming==0.13.3
Pillow==10.1.0
psycopg2==2.8.6
psycopg2-binary==2.9.9
pycodestyle==2.10.0
pycparser==2.20
pyflakes==3.0.1
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2021.3
redis==5.0.1
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==1.39.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
yookassa==2.3.5
zipp==3.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.14
django-redis==5.4.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-cleanup==8.0.0
django-ratelimit==4.1.0
django-cacheops==6.1.1
django-prometheus==2.3.1
prometheus-client==0.19.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.6.0
coverage==7.4.1
